Add GeneralMeasureTasks to hold generic measurements
